Add StartPage component and test for scrollable faction list
