+++ src/__tests__/rift.ui.spec.tsx
@@ -1,13 +1,13 @@
 import { describe, it, expect } from 'vitest'
 import { render, screen } from '@testing-library/react'
 import { CompactShip } from '../components/ui'
-import { makeShip, getFrame, PARTS, type Part } from '../game'
+import { makeShip, getFrame, PARTS, RARE_PARTS, type Part } from '../game'
 
 describe('CompactShip rift dice display', () => {
   it('shows rift dice when only rift conductor is installed', () => {
     const src = PARTS.sources[0]
     const drv = PARTS.drives[0]
-    const conductor = PARTS.hull.find(p=>p.id==='rift_conductor')!
+    const conductor = RARE_PARTS.find(p=>p.id==='rift_conductor')!
     const ship = makeShip(getFrame('interceptor'), [src, drv, conductor])
     render(<CompactShip ship={ship} side="P" active={false} />)
     expect(screen.queryByText(/No weapons/i)).toBeNull()
